+++ ensek/client.py
@@ -1,4 +1,5 @@
 import logging
+from urllib.parse import urljoin
 from http.client import NOT_FOUND, INTERNAL_SERVER_ERROR, BAD_REQUEST
 from string import Template
 
@@ -19,8 +20,12 @@ class Ensek:
 
     ENDPOINTS = {
         'get_account': Template('/accounts/$account_id'),
+        'get_meter_point_readings': Template(
+            '/MeterPoints/$meter_point_id/Readings'
+        ),
         'get_completed_signups': Template('/SignUps/Completed'),
         'get_meter_points': Template('/Accounts/$account_id/MeterPoints'),
+        'create_meter_reading': Template('/Accounts/$account_id/Readings'),
         'get_region_id_for_postcode': Template('/Regions/$postcode'),
         'get_gas_utility': Template('/UtilitiesLookups/Gas/$mprn'),
         'get_electricity_utility': (
@@ -38,10 +43,41 @@ class Ensek:
         self._resource_path = None
         self._headers = {'Authorization': f'Bearer {self._api_key}'}
 
-    def _get(self, path, params):
-        url = f'{self._api_url}/{path.lstrip("/")}'
+    def create_meter_reading(
+        self, *, account_id, meter_point_id, register_id, value, timestamp
+    ):
+        path = self.ENDPOINTS['create_meter_reading'].substitute(
+            account_id=account_id
+        )
+        body = [
+            {
+                'meterPointId': int(meter_point_id),
+                'dateTime': timestamp.isoformat(),
+                'readings': [{
+                    'registerId': int(register_id),
+                    'value': float(value),
+                }],
+            }
+        ]
+        return self._post(path=path, body=body)
+
+    def _path_to_full_url(self, path):
+        return urljoin(self._api_url, path.lstrip('/'))
+
+    def _get(self, path, params=None):
+        url = self._path_to_full_url(path)
         try:
             response = requests.get(url, headers=self._headers, params=params)
+        except RequestException as exc:
+            raise EnsekError(exc, response=None) from exc
+        if not response.ok:
+            self._handle_bad_response(response)
+        return response.json()
+
+    def _post(self, *, path, body):
+        url = self._path_to_full_url(path)
+        try:
+            response = requests.post(url, headers=self._headers, json=body)
         except RequestException as exc:
             raise EnsekError(exc, response=None) from exc
         if not response.ok:
@@ -62,9 +98,14 @@ class Ensek:
             raise EnsekError(msg, response=response)
 
     def __getattr__(self, name):
-        try:
-            self._resource_path = self.ENDPOINTS[name]
-        except KeyError:
+        if name.startswith('get_'):
+            try:
+                self._resource_path = self.ENDPOINTS[name]
+            except KeyError:
+                return self.__dict__[name]
+        else:
+            # If the client method doesn't have `get_` prefix, assume it's
+            # defined on the class
             return self.__dict__[name]
         return self